refactor: service layer — extract CurrentUserService, UrlFetchService, and make username immutable
- Extract CurrentUserService to eliminate duplicated getCurrentUser() logic
- Extract UrlFetchService from AiExtractionService (SRP: SSRF protection + URL fetching)
- Make username immutable — remove from UpdateProfileRequest (backend + frontend)
- Set statusChangedDate on creation and on actual status changes
- Remove username-related controller tests
- Add 24 UrlFetchService tests for SSRF protection
