+++ backend/src/main/java/com/nadavramon/job_tracker/service/CurrentUserService.java
@@ -1,0 +1,23 @@
+package com.nadavramon.job_tracker.service;
+
+import com.nadavramon.job_tracker.entity.User;
+import com.nadavramon.job_tracker.exception.ResourceNotFoundException;
+import com.nadavramon.job_tracker.repository.UserRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CurrentUserService {
+
+    private final UserRepository userRepository;
+
+    public CurrentUserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    public User getCurrentUser() {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+    }
+}
